blog-automation/backend/requirements.txt
# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Celery
celery==5.3.4
flower==2.0.1

# AI APIs
openai==1.3.5
anthropic==0.7.7

# Web scraping and automation
playwright==1.40.0
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1

# SEO and analytics
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Utils
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2
structlog==23.2.0
prometheus-client==0.19.0
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5